+++ tests/screenshot/apps/LimestoneComponents.js
@@ -6,6 +6,7 @@ import Alert from './components/Alert';
 import BodyText from './components/BodyText';
 import Button from './components/Button';
 import Card from './components/Card';
+import CardSports from './components/CardSports';
 import Checkbox from './components/Checkbox';
 import CheckboxItem from './components/CheckboxItem';
 import Chip from './components/Chip';
@@ -64,6 +65,7 @@ const components = {
 	BodyText,
 	Button,
 	Card,
+	CardSports,
 	Checkbox,
 	CheckboxItem,
 	Chip,